requirements.txt
Django==4.1.6
djangorestframework==3.14.0
pytz==2018.9
sqlparse==0.3.0
django-environ==0.9.0
psycopg2-binary==2.9.5
psycopg2==2.9.5
django-postgres-extra==2.0.8
django-extensions==3.2.1
django-bootstrap5==22.2
pdfkit==1.0.0
XlsxWriter==3.0.8
django-prometheus==2.2.0
prometheus-client==0.16.0
requests==2.28.2
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
urllib3==1.26.15